\name{DrawMargin}
\alias{DrawMargin}
\title{Compute EPSILON or draw margin, a measure of how likely a draw is.
takes draw probability as input
Can use draw probability to estimate draw margin
or vice versa}
\usage{
  DrawMargin(draw_probability = 0.5, beta = 1,
    total_players = 2)
}
\arguments{
  \item{draw_probability}{draw probability}

  \item{beta}{randomness in game}

  \item{total_players}{number of players}
}
\description{
  Compute EPSILON or draw margin, a measure of how likely a
  draw is. takes draw probability as input Can use draw
  probability to estimate draw margin or vice versa
}

