\name{DrawProbability}
\alias{DrawProbability}
\title{Compute the draw probability given the draw margin (epsilon)
Can use draw probability to estimate draw margin
or vice versa}
\usage{
  DrawProbability(epsilon = 0.1, beta = 1,
    total_players = 2)
}
\arguments{
  \item{epsilon}{how common draws area}

  \item{beta}{randomness in game}

  \item{total_players}{number of players}
}
\description{
  Compute the draw probability given the draw margin
  (epsilon) Can use draw probability to estimate draw
  margin or vice versa
}

