\docType{data}
\name{data}
\alias{data}
\title{Australian Open data}
\format{A data frame with 127 rows and 5 variables}
\description{
  A dataset containing tennis players and match outcomes
  from the 2012 Australian Open. Obtained from \url{http://www.tennis-data.co.uk/data.php} (2013).
}
\section{trueskill-package:}{
Refer to \code{\link{trueskill-package}} for links to related classes, functions and data.
}
\details{
  \itemize{ 
    \item Winner Player that won the game 
    \item Loser Player that lost the game 
    \item Round Round of the tournament 
    \item WRank World Tennis Association Ranking of winning player 
    \item LRank World Tennis Association Ranking of losing player }
}
\keyword{datasets}

